refactor: Round 관련 상수 및 타입 정리
- constants 폴더에 있는 'miniGame.ts'을 types 폴더의 'round.ts'로 이동
- 'CardGamePlayPage'에서 타이머 관련 상수 수정
  - TOTAL_COUNT가 아닌, ROUND_TOTAL_TIME으로 네이밍 변경 및 Props로 내려주도록 수정함
  - 따로 상수 파일을 사용할 정도로 많은 파일에서 사용되는 것도 아니고, CardGamePlayPage에서 Round로 내려주기만 하면 됨
